addition to r1121: now fixed on windows too: inner files (files listed in nzb) bigger than 2GB could not be downloaded

+++ trunk/daemon/main/Options.cpp
@@ -3013,7 +3013,7 @@ bool Options::SaveConfig(OptEntries* pOptEntries)
 	}
 
 	// close and truncate the file
-	int pos = ftell(infile);
+	int pos = (int)ftell(infile);
 	fclose(infile);
 
 	Util::TruncateFile(m_szConfigFilename, pos);
